Refactor sitecontent response-funksjoner for å skille tydeligere mellom draft og master requests
